pyproject.toml
[project]
name = "emotions"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.129.0",
    "google-genai>=1.63.0",
    "python-dotenv>=1.2.1",
    "streamlit>=1.54.0",
    "uvicorn>=0.40.0",
]
